Module tink_signature::subtle

This module provides subtle implementations of digital signature primitives.

Structs

EcdsaSigner
EcdsaSigner is an implementation of tink::Signer for ECDSA. At the moment, the implementation only accepts DER encoding.

EcdsaVerifier

Ed25519Signer
A Signer implementation for ED25519.

Ed25519Verifier
A tink::Verifier for ED25519.

Enums

EcdsaPrivateKey

EcdsaPublicKey
An ECDSA public key.

SignatureEncoding
Supported signature encodings. This is a precise subset of the protobuf enum, allowing exact matches.

Functions

validate_ecdsa_params
Validate ECDSA parameters. The hash's strength must not be weaker than the curve's strength. Only DER encoding is supported now.